NB-131 : feat - 인증 기반 게시글 삭제 기능 추가
- 유저 소유의 게시글 검증 기능
- Forbidden 예외처리 클래스 생성

+++ src/main/java/com/soyeon/nubim/domain/post/PostControllerV1.java
@@ -83,10 +83,14 @@ public class PostControllerV1 {
 		}
 		return ResponseEntity.ok(postService.findAllPostsByUserIdOrderByCreatedAt(userId, pageRequest));
 	}
-	
+
 	@DeleteMapping("{postId}")
 	public ResponseEntity<Void> deletePost(@PathVariable Long postId) {
+		postService.validatePostExist(postId);
+		postService.validatePostOwner(postId, userService.getCurrentUser());
+
 		postService.deleteById(postId);
+
 		return ResponseEntity.ok().build();
 	}
